+++ src/components/taskManagerFolders/tasks/forms/EditTasksForm.jsx
@@ -1,8 +1,8 @@
 import { useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 
-import { updateTask, resetTaskStatus } from "../../../redux/tasksSlice";
-import { toastWithProgress } from "../utils/toastWithProgress";
+import { updateTask, resetTaskStatus } from "../../../../redux/tasksSlice";
+import { toastWithProgress } from "../../utils/toastWithProgress";
 import TaskFormFields from "./TaskFormFields";
 
 export default function EditTaskForm({ task, onClose }) {
